fullStack/backend/requirements.txt
fastapi==0.115.0
numpy==2.1.1
protobuf==5.28.2
pydantic==2.9.2
pymongo==3.10.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
xgboost==2.1.1
